+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/mojochao/devbox-cli/cmd"
+import "github.com/mojochao/devbox/cmd"
 
 func main() {
   cmd.Execute()