+++ RELEASES.md
@@ -1,5 +1,11 @@
 # RELEASES
 
+## 0.2.0
+
+- Improved docs
+- Improved flexibility of start, stop, and remove commands
+- Renamed go module to just devbox instead of devbox-cli
+
 ## 0.1.0
 
 - Initial implementation